backend/pyproject.toml
[tool.poetry]
name = "rag-financial-backend"
version = "1.0.0"
description = "RAG Financial POC Backend - Exercise 3: Adaptive Knowledge Base Management"
authors = ["Devin AI <devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
# Azure Services
azure-search-documents = "^11.4.0"
azure-cosmos = "^4.5.1"
azure-ai-formrecognizer = "^3.3.2"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.7.0"
azure-monitor-opentelemetry = "^1.2.0"
azure-ai-ml = "^1.12.0"
azure-ai-generative = "^1.0.0b3"
# OpenAI and AI
openai = "^1.12.0"
semantic-kernel = "^0.9.6b1"
# FastAPI and Web
fastapi = {extras = ["standard"], version = "^0.109.2"}
uvicorn = {extras = ["standard"], version = "^0.29.0"}
python-multipart = "^0.0.9"
httpx = "^0.26.0"
# Data Processing
pandas = "^2.2.0"
numpy = "^1.26.4"
python-docx = "^1.1.0"
PyPDF2 = "^3.0.1"
openpyxl = "^3.1.2"
tiktoken = "^0.5.2"
# Observability and Monitoring
opentelemetry-api = "^1.22.0"
opentelemetry-sdk = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-instrumentation-httpx = "^0.43b0"
psutil = "^5.9.8"
# Additional dependencies
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"
msal = "^1.25.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "uvicorn app.main:app --host 0.0.0.0 --port 8000"
